backend/requirements.txt
python-binance
pandas
numpy
pyyaml
python-dotenv
matplotlib